requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
opencv-python-headless==4.10.0.84
google-generativeai==0.8.3
python-dotenv==1.0.1
pyserial==3.5
httpx==0.27.0
websockets==13.0
Pillow==10.4.0
numpy==1.26.4
